Trivial: Prevent UB in class CNoDestination
This commit fixes a weird/UB issue in class `CNoDestination`, which was
allowing for `a < b` and `a == b` to both be true at the same time.

The `CNoDestination` type is used to indicate non-standard scripts that
don't match known address patterns. This class is conceptually the same
as `std::monostate` and all instances should compare to being equal.

Thus, `operator<` returning `true` is nonsensical and may cause UB.

This is because in C++ std library, no more than one of these may be
true for a given type: `a < b`, `a == b`, `a > b`.

Otherwise you may get UB if trying to use a class that violates the
above predicate in a std C++ container.