Remove boost::variant dependency, use std::variant instead.
This commit removes the dependency/usage in the codebase of all
boost::variant usages and replaces them with std::variant.

Some of the visitor functor objects were also replaced with lambdas,
using the "Overloaded" idiom.

Also in this commit: some slight refactoring in key_io.cpp where the
static function `DecodeLegacyDestination` which was used in only 1
place, in `DecodeLegacyAddr`, was simply rolled-into `DecodeLegacyAddr`.