Guard against unsafe PSBT tx usage
In this codebase, there are some alarmingly unsafe (potential for UB)
usages of the `PartiallySignedTransaction::tx` public member. This
member is a `std::optional` and dereferencing it unconditionally with
`*psbtx.tx` could lead to C++ UB.

Thankfully, in all current codepaths this does not occur, but the
`PartiallySignedTransaction` class itself as designed is very unsafe
to use. Namely, if you default-construct it as
`PartiallySignedTransaction{}`, and then try to serialize that empty tx,
you get UB!

This commit fixes the situation by guarding all unconditional
dereferences of `*psbtx.tx` with either the `std::optional` `.value()`
usage or with our own helper macro `CHECK_NONFATAL`, both of which throw
exceptions in the `!psbtx.has_value()` case.

This should at least guarantee no UB in the app from invalid PSBT usage, even
if  it can lead to exceptions being thrown on invalid usage.

Since UB is always terrible to have in a C++ app, the guards/checks that
may throw exceptions are always preferable to the current potential for
UB.