Fix for failure to compile UniValue using older GCC 8.3.0
Summary
---

When doing the gitian build against master, I observed it fails to build
as of the merge of !1922.

This is because we use an older GCC 8.3.0 in the gitian builds and this
compiler version has a bug. It complains as follows:

    ../src/univalue/include/univalue.h:839:27: error: deduction guide 'UniValue::Overloaded(Ts ...) -> UniValue::Overloaded<Ts ...>' must be declared at namespace scope
         template<class... Ts> Overloaded(Ts...) -> Overloaded<Ts...>;

This is an incorrect error as per the C++17 standard, so it's a compiler
bug.

The fix is to take this templated class out of the class it nests in and
add a header `"univalue_detail.h"` that does the same thing, but at
namespace-scope level.

Test Plan
---

- `ninja all check`
- OPTIONAL: Observe the error by doing a gitian-build vs master which
  should fail. Then do a gitian-build vs this MR and it should succeed.