Fix for CoinStatsIndex compiler bug (gcc 8.3.0)
Summary
---

On GCC 8.3.0, a nonsensical error is emitted when compiling the
codebase.

    In file included from /usr/include/c++/8/memory:80,
                     from ../src/./serialize.h:21,
                     from ../src/./amount.h:9,
                     from ../src/./rpc/blockchain.h:7,
                     from ../src/rpc/blockchain.cpp:7:
    /usr/include/c++/8/bits/unique_ptr.h: In instantiation of 'void std::default_delete<_Tp>::operator()(_Tp*) const [with _Tp = CoinStatsIndex::BestBlockStats]':
    /usr/include/c++/8/bits/unique_ptr.h:274:17:   required from 'std::unique_ptr<_Tp, _Dp>::~unique_ptr() [with _Tp = CoinStatsIndex::BestBlockStats; _Dp = std::default_delete<CoinStatsIndex::BestBlockStats>]'
    ../src/./index/coinstatsindex.h:29:87:   required from here
    /usr/include/c++/8/bits/unique_ptr.h:79:16: error: invalid application of 'sizeof' to incomplete type 'CoinStatsIndex::BestBlockStats'
      static_assert(sizeof(_Tp)>0,
                    ^~~~~~~~~~~

This is due to incorrect implementation of `std::unique_ptr` where it should be
usable with opaque (incomplete) types in this context. However the
compiler barfs here, because it doesn't like the incomplete type.  On  other compilers
it works and the C++ standard says it should work.

The fix is to make the member `CoinStatsIndex::m_best_block_stats` a `std::shared_ptr`
instead which works ok here for incomplete types on GCC 8.3.0 (and all
other compilers).

Test Plan
---

- `ninja all check`
- OPTIONAL: See if you can use gcc 8.3.0 to reproduce the problem, then
  use this commit and see the problem disappear.