Fix for clang-8 compile issue
Summary
---

When compiling for OSX using gitian, I got the following (erroneous)
error:

    In file included from ../src/init.cpp:31:
    In file included from ../src/./index/coinstatsindex.h:9:
    ../src/./coinstats.h:59:5: error: exception specification of explicitly defaulted default constructor does not match the calculated one
        CoinStatsBase() noexcept = default;
        ^
    ../src/./coinstats.h:83:5: error: exception specification of explicitly defaulted default constructor does not match the calculated one
        CoinStats() noexcept = default;
        ^

The fact that this produces an error deviates from the latest C++
standard (maybe fixed in a defects report?).

Anyway clang-8 which is what is used for the OSX gitian build fails.
This commit fixes the situation by not using an `= default` constructor,
just use an empty one and the error goes away.

Test Plan
---

- `ninja all check`
- OPTIONAL: Use clang-8 (or gitian osx build) and reproduce the problem
  and observe it being fixed by this commit.