Fix failing uint256_tests if compiled with gcc 15.1.1 (arch linux)
The test was failing due to use of C++ UB, which happened to work ok on
previous clang and gcc but latest GCC 15.1.1 on Arch Linux compiles the
code but the test fails.

The key to why it was failing: you aren't allowed to rely on the particular
contents of uninitialized memory.

Also we were re-using this memory after a placement-new operation, which is
UB unless you std::launder. The fix involves re-doing the logic of the test to:

1. create the uninitialized object on top of alignedPtr
2. check its data pointer == alignedPtr
3. destruct the uninitialized object
4. Launder alignedPtr so that it is "usable" again without UB.
5. Check contents of the memory.