Make CBlockIndex::nChainTx be a 64-bit unsigned int
Made the memory-only member `nChainTx` of class `CBlockIndex` be 64-bit
rather than 32-bit. Right now the chain is at ~400 million txns but
since we are a large-block chain, we may end up with >4 billion txns on
the chain sooner than the comment said (2030 would be for BTC, not for
us).

This change future-proofs `CBlockIndex` for big blocks, at a cost of 4
extra bytes per header we track in memory.

The `nChainTx` member is a derived value and is re-computed each time at
node startup and is not persisted to disk.

Note that the member function `CBlockIndex::GetChainTxCount()` was also
modified to return an unsigned 64-bit value rather than the signed
64-bit value it was returning previously, since the signed value makes
no sense (one cannot have negative nChainTx!).

Also note that `nChainTx` member was moved down in the class in order to
avoid wasteful padding bytes.

These changes only affect logging and the rpc `getchaintxstats`, and are
not at all consensus-critical.