Use std::move in unserialization of std::set and std::map
This is a minor performance improvement for the unserialization of
std::map and std::set types. Previous code was not using std::move to
insert the temporary unserialized object into the map/set. This has been
addressed and should offer a speedup when unserializing types that
benefit from move-semantics. Other types are unaffected.

Also in this commit: A small nit to prefer ++i to i++ and to use `const`
and `auto` to improve readability.