Optimization: make use of CTOR when searching for txn in a block
- IF the block is on the active chain, and 
- IF it's beyond "magnetic anomaly" height (CTOR activation), and 
- IF the block is sufficiently large, then

It makes sense to use a binary search to find (by txid) a particular 
transaction in a block, since we know the block is sorted. If all 3
of the above conditions are not met, then we fall back to linear search.

Several call sites that utilize lookup by txid were modified and code
was carefully modified to ensure the above 3 preconditions hold for
the binary search to be allowed to proceed.

Co-authored-by: Calin Culianu <calin.culianu@gmail.com>