refactor: Move most CAutoFile methods into an implementation file
We move most of the code from class CAutoFile declared in streams.h into
a (newly created) implementation file, streams.cpp

This simplifies streams.h by not having non-template code appear in the
file, thus making compilation times potentially faster.

The motivation for this change is to synchronize the codebase better
with Core (which also has done this change), so as to make backports
easier.