Slightly modernize the code in blockstorage and add overflow checks
We add overflow checking to some functions in `blockstorage.cpp` related
to undo and block file reads/writes. This is to detect at runtime if app
invariants are violated (such that at the present time blocks and block
files may not exceed 4GiB).

Also in this commit we switch calls to bare `ftell()` to the more
C++-standard `std::ftell()`.

This is a pure code quality commit.