Make prevector class convertible to std::span
The reason it wasn't was because the `iterator` type it implemented was
not tagged as being a contiguous_iterator (required in C++20 to be
convertible to std::span).

We could have tagged the `prevector::iterator` type appropriately, but
there is an easier way.

It turns out that the easiest way to get this is just to make the
iterator type be a raw pointer. There is no disadvantage in doing that
and it allows us to delete a bunch of unneeded code in the prevector
class after doing that.

Also in this commit: Modernize some typedefs in the prevector class and
also added a unit test that tests that the std::span implicit conversion
works and is sane.