Added support for serialization of std::span<uint8_t>
Serialization support for `std::span` byte buffers may be needed in future
work. As such, we add support for it. The semantics are identical to our
custom `Span` type, where the raw bytes are written & read without a
leading compact size (contrast this to `std::vector` which has a leading
compact size prepended to the serialized stream).

Also in this commit: Added a unit test to `serialize_tests.cpp` that
tests `Span<uint8_t>` and `std::span<uint8_t>` serialization behaves as
expected.