util: Filter control characters out of log messages
Co-Authored-By: laanwj <126646+laanwj@users.noreply.github.com>

This is a heavily modified backport of Core: https://github.com/bitcoin/bitcoin/pull/17095

We modified the Core implementation to be more performant and not do any
allocations or copying in the likely "happy path" of not needing to
escape the log string (this is the case 99.9999% of the time).

Contrast this to the Core version which always would allocate and copy, even in the
common case.

Otherwise our implementation is functionally identical.

Original commit message:

    Belts and suspenders: make sure outgoing log messages don't contain
    potentially suspicious characters, such as terminal control codes.

    This escapes control characters except newline ('\n') in C syntax.
    It escapes instead of removes them to still allow for troubleshooting
    issues where they accidentally end up in strings.