logging: Added unit tests for the new rate limiting facility
This is a commit that is heavily modified from Core#32604, and also
fast-forwarded to latest Core and also pulls in other pieces we will
need to add the logging rate limiter unit test.

Creates a new file called: logging_tests.cpp, which currently only
contains unit tests for the rate limiter facility. But more logging
tests can be added to this file in the future.

Co-Authored-By: Eugene Siegel <elzeigel@gmail.com>
Co-Authored-By: Niklas Gogge <n.goeggi@gmail.com>
Co-Authored-By: stickies-v <stickies-v@protonmail.com>