gitian: Remove installation of ALL of boost in favor of the pieces we use
In this commit, we simply explicitly install just the pieces of boost
that our codebase uses in the gitian container's host system, rather
than libboost-all-dev which installs all of boost.

This is in the spirit of minimalism and efficiency and to make
setup/installation of the gitian container image slightly faster.