Update checkpoints for v29.0.0 release
Summary
---

The checkpoints for mainnet, testnet3, testnet4, and chipnet are updated
by this commit.  They are all >= 10 blocks in the past for mainnet,
>=2000 blocks in the past for the testnet(s),

Test Plan
---

Use a block explorer e.g. https://bch.loping.net to verify the heights
and hash match for each of the 4 entries added for each respective chain, and
that the hash and height are on the currently active best chain (should
have confirmations: xxx, >= 11 for mainnet, >= 2001 for testnets).